Доработан протокол Sony для чтения команд длинной 12, 15 и 20 бит Исправлена ошибка в библиотеке IRLib, которая могла помешать передавать сигнал с паузами в конце

+++ libs/IRLib/IRLib.h
@@ -45,7 +45,7 @@ class IRLib {
 		void			ExternFillPostOperations();
 
 		void			AppendRawSignal(IRLib &);
-		int16_t			RawPopItem();
+		int32_t			RawPopItem();
 
 		static bool		CompareIsIdentical(IRLib &Signal1, IRLib &Signal2);
 	private:
